enable R2 tests to run in AAT
