--- src/test/functionalTests/tests/nonprod/LipvLip_UI_DefaultJudgment_tests.js
@@ -1,23 +1,0 @@
-const config = require('../../../config');
-
-const LoginSteps = require('../../commonFeatures/home/steps/login');
-const {createAccount} = require('../../specClaimHelpers/api/idamHelper');
-const ClaimantResponseSteps = require('../../citizenFeatures/response/steps/lipClaimantResponseSteps');
-
-const claimType = 'SmallClaims';
-// eslint-disable-next-line no-unused-vars
-let claimRef;
-
-Feature('Create Lip v Lip claim -  Default Judgment');
-
-Scenario('Create LipvLip claim and defendant not responded by deadline and Claimant raise Default Judgment', async ({api}) => {
-  if (['preview', 'demo'].includes(config.runningEnv)) {
-    await createAccount(config.claimantCitizenUser.email, config.claimantCitizenUser.password);
-    await createAccount(config.defendantCitizenUser.email, config.defendantCitizenUser.password);
-    claimRef = await api.createLiPClaim(config.claimantCitizenUser, claimType);
-    await api.amendRespondent1ResponseDeadline(config.systemUpdate2);
-    await LoginSteps.EnterCitizenCredentials(config.claimantCitizenUser.email, config.claimantCitizenUser.password);
-    await ClaimantResponseSteps.verifyDefaultJudgment(claimRef);
-    await api.waitForFinishedBusinessProcess();
-  }
-}).tag('@skip-regression-r2');
